# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GinWeb脚手架项目",
        "title": "GinWeb脚手架项目",
        "contact": {
            "name": "八宝糖",
            "email": "1013269096@qq.com"
        },
        "version": "v1.0.0"
    },
    "host": "127.0.0.1",
    "basePath": "/api/v1/",
    "paths": {
        "/login": {
            "post": {
                "description": "处理用户登录流程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "用户登录接口",
                "parameters": [
                    {
                        "description": "用户登录所需参数",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.ParamsLogin"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/signup": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册所需参数",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.ParamsSignUp"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.ParamsLogin": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.ParamsSignUp": {
            "type": "object",
            "required": [
                "email",
                "password",
                "repassword",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "repassword": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}